Update to Oublek's dialog
Update to Oublek's reaction if PC impersonated Greywolf for payment
from Oublek. If reaction is HOSTILE, Oublek refuses to deal with the PC
and EscapeArea(). This makes Oublek's line and the journal entry make
sense. If the reaction is NEUTRAL or FRIENDLY, Oublek chastises the PC
(with unused line of Oublek dialog, #665), but will accept bounties (at
reduced rewards).