Fix conflict with BGFixPack Game Text Update
BGFixpack changes the name of the tavern by the bridge in Nashkel to "Belching Dragon" by rewriting string #11696 in the dialog.tlk. BG1UB points the name for the "Nashkel Inn" to that string, assuming it is still "Northern Light", as per the original dialog.tlk.

Preventing renaming of Nashkel Inn if BGFixPack Game Text Update is installed.

http://gibberlings3.net/forums/index.php?showtopic=29374&p=260589